# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: zboard API
  version: v0.0.1
  description: |
    zboard single-binary operations console API for subscriber and operator workflows.
    Business APIs use /api/v1. Zero's standard event webhook is exposed at /api/zero/events.
servers:
  - url: http://localhost:8080
  - url: https://api.example.com
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/admin/subscription-templates:
    get:
      tags: [subscriptions, admin]
      summary: List subscription export templates
      description: Set paged=true for the canonical page envelope and summary rows without template source. The unpaged legacy response includes source for compatibility.
      parameters:
        - { name: paged, in: query, schema: { type: boolean }, description: Return the canonical page envelope with summary rows. }
        - { name: q, in: query, schema: { type: string, maxLength: 100 } }
        - { name: active, in: query, schema: { type: boolean } }
        - { name: offset, in: query, schema: { type: integer, minimum: 0, default: 0 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
      responses:
        "200":
          description: Subscription template list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - properties:
                      data:
                        oneOf:
                          - type: array
                            items: { $ref: "#/components/schemas/SubscriptionTemplate" }
                          - $ref: "#/components/schemas/SubscriptionTemplatePage"
    post:
      tags: [subscriptions, admin]
      summary: Create and validate a subscription export template
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SubscriptionTemplateWriteRequest" }
      responses:
        "200": { description: Template created }
        "400": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    NodeConnectorBearer:
      type: http
      scheme: bearer
      bearerFormat: ZeroNodeAPIKey
  responses:
    Error:
      description: Request failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiResponse"
  schemas:
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: object
          nullable: true
        error:
          $ref: "#/components/schemas/ApiErrorDetail"
        timestamp:
          type: string
    ApiErrorDetail:
      type: object
      description: Versioned machine-readable detail included on HTTP error responses. The top-level message remains for backward compatibility.
      required: [version, code]
      properties:
        version:
          type: integer
          enum: [1]
        code:
          type: string
          example: validation_failed
        fields:
          type: object
          description: Field-path to localized actionable message. Present when code is validation_failed.
          additionalProperties:
            type: string
    PageMetadata:
      type: object
      required: [offset, limit, total, next_cursor, previous_cursor]
      properties:
        offset: { type: integer, minimum: 0 }
        limit: { type: integer, minimum: 1, maximum: 200 }
        total: { type: integer, format: int64, minimum: 0 }
        next_cursor: { type: string, nullable: true }
        previous_cursor: { type: string, nullable: true }
    PageEnvelope:
      type: object
      required: [page, aggregates, facets]
      properties:
        page: { $ref: "#/components/schemas/PageMetadata" }
        aggregates: { type: object, additionalProperties: true }
        facets: { type: object, additionalProperties: true }
        total: { type: integer, format: int64, minimum: 0, deprecated: true, description: Compatibility alias for page.total. }
        offset: { type: integer, minimum: 0, deprecated: true, description: Compatibility alias for page.offset. }
        limit: { type: integer, minimum: 1, maximum: 200, deprecated: true, description: Compatibility alias for page.limit. }
...
    SubscriptionTemplate:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string, maxLength: 80 }
        slug: { type: string, maxLength: 80, pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$" }
        description: { type: string, maxLength: 255 }
        renderer: { type: string, enum: [znet-sink, clash, sing-box, unsupported], description: Backend-owned base encoder. Unsupported is emitted only for disabled legacy records awaiting conversion. }
        customization:
          $ref: "#/components/schemas/SubscriptionTemplateCustomization"
        content_type: { type: string, enum: [text/plain, text/yaml, application/yaml, application/json], description: Read-only response type derived from the renderer. Omitted for unsupported legacy records. }
        is_active: { type: boolean }
        sort_order: { type: integer }
        revision: { type: integer, format: int64, minimum: 1 }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    SubscriptionTemplatePage:
      allOf:
        - $ref: "#/components/schemas/PageEnvelope"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: "#/components/schemas/SubscriptionTemplate" }
    SubscriptionTemplateWriteRequest:
      type: object
      required: [name, slug, renderer, is_active]
      properties:
        name: { type: string, minLength: 1, maxLength: 80 }
        slug: { type: string, maxLength: 80, pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$" }
        description: { type: string, maxLength: 255 }
        renderer: { type: string, enum: [znet-sink, clash, sing-box] }
        customization:
          $ref: "#/components/schemas/SubscriptionTemplateCustomization"
        is_active: { type: boolean }
        sort_order: { type: integer }
        expected_revision: { type: integer, format: int64, minimum: 1, description: Optional optimistic concurrency guard for updates. }
...
    SubscriptionTemplateCustomization:
      type: object
      additionalProperties: false
      description: Declarative, non-executable policy groups and rule bindings applied to a backend-owned renderer. Dynamic nodes and credentials remain backend-managed.
      required: [version, main_group, policy_groups, final, rule_sets]
      properties:
        version: { type: integer, enum: [2] }
        main_group: { type: string, pattern: "^[a-z][a-z0-9-]{0,31}$", description: ID of the policy group used as the primary client selector. }
        policy_groups:
          type: array
          minItems: 1
          maxItems: 16
          items:
            $ref: "#/components/schemas/SubscriptionPolicyGroup"
        final: { type: string, pattern: "^(direct|group:[a-z][a-z0-9-]{0,31})$", description: Final routing target. }
        rule_sets:
          type: array
          maxItems: 64
          items:
            $ref: "#/components/schemas/SubscriptionTemplateRuleSet"
        advanced_source:
          type: string
          maxLength: 131072
          description: YAML for Clash or JSON for ZNet Sink and sing-box. Deep-merged after visual configuration and validated as a final client document. Root proxies/outbounds replacement must retain the documented Zboard generated injection marker; $zboard:all-nodes expands node names in member arrays.
    SubscriptionPolicyGroup:
      type: object
      additionalProperties: false
      required: [id, name, type]
      properties:
        id: { type: string, pattern: "^[a-z][a-z0-9-]{0,31}$" }
        name: { type: string, minLength: 1, maxLength: 64, description: Native client policy-group name or tag. }
        type: { type: string, enum: [select, urltest, fallback], description: fallback is supported only by Clash and ZNet Sink. }
        include_pattern: { type: string, maxLength: 256, description: Optional RE2 regular expression matched against protocol configuration names. Empty includes every available node. }
        exclude_pattern: { type: string, maxLength: 256, description: Optional RE2 regular expression applied after include_pattern. }
        include_groups:
          type: array
          maxItems: 15
          uniqueItems: true
          items: { type: string, pattern: "^[a-z][a-z0-9-]{0,31}$" }
        default_group: { type: string, pattern: "^[a-z][a-z0-9-]{0,31}$", description: Optional included group selected first by a select group. }
        probe_url: { type: string, format: uri, maxLength: 2048 }
        interval: { type: integer, minimum: 60, maximum: 86400 }
        tolerance: { type: integer, minimum: 0, maximum: 10000 }
    SubscriptionTemplateRuleSet:
      type: object
      additionalProperties: false
      required: [target]
      description: Either a reusable rule-set reference with rule_set_id, or a template-only remote shortcut with source fields.
      oneOf:
        - required: [rule_set_id, target]
        - required: [tag, url, format, target, interval]
      properties:
        rule_set_id: { type: integer, format: int64, minimum: 1, description: Reusable rule-set ID. Source fields cannot override a referenced record. }
        tag: { type: string, minLength: 1, maxLength: 64, pattern: "^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$" }
        url: { type: string, format: uri, maxLength: 2048 }
        behavior: { type: string, enum: [domain, ipcidr, classical], description: Clash-only provider behavior. }
        format: { type: string, enum: [domain_list, cidr_list, zero_rule_ir, zrs, yaml, text, mrs, source, binary] }
        target: { type: string, pattern: "^(direct|reject|group:[a-z][a-z0-9-]{0,31})$", description: Policy-group ID, direct or reject target used when this rule set matches. }
        interval: { type: integer, minimum: 60, maximum: 604800 }
...